reimplement `eliminate_outer_join`
