+++ datafusion/optimizer/src/lib.rs
@@ -21,6 +21,7 @@ pub mod decorrelate_where_in;
 pub mod eliminate_cross_join;
 pub mod eliminate_filter;
 pub mod eliminate_limit;
+pub mod eliminate_outer_join;
 pub mod filter_null_join_keys;
 pub mod filter_push_down;
 pub mod inline_table_scan;
@@ -28,7 +29,6 @@ pub mod limit_push_down;
 pub mod optimizer;
 pub mod projection_push_down;
 pub mod propagate_empty_relation;
-pub mod reduce_outer_join;
 pub mod scalar_subquery_to_join;
 pub mod simplify_expressions;
 pub mod single_distinct_to_groupby;
