+++ datafusion/optimizer/src/optimizer.rs
@@ -23,13 +23,13 @@ use crate::decorrelate_where_in::DecorrelateWhereIn;
 use crate::eliminate_cross_join::ReduceCrossJoin;
 use crate::eliminate_filter::EliminateFilter;
 use crate::eliminate_limit::EliminateLimit;
+use crate::eliminate_outer_join::EliminateOuterJoin;
 use crate::filter_null_join_keys::FilterNullJoinKeys;
 use crate::filter_push_down::FilterPushDown;
 use crate::inline_table_scan::InlineTableScan;
 use crate::limit_push_down::LimitPushDown;
 use crate::projection_push_down::ProjectionPushDown;
 use crate::propagate_empty_relation::PropagateEmptyRelation;
-use crate::reduce_outer_join::ReduceOuterJoin;
 use crate::rewrite_disjunctive_predicate::RewriteDisjunctivePredicate;
 use crate::scalar_subquery_to_join::ScalarSubqueryToJoin;
 use crate::simplify_expressions::SimplifyExpressions;
@@ -183,7 +183,7 @@ impl Optimizer {
         if config.filter_null_keys {
             rules.push(Arc::new(FilterNullJoinKeys::default()));
         }
-        rules.push(Arc::new(ReduceOuterJoin::new()));
+        rules.push(Arc::new(EliminateOuterJoin::new()));
         rules.push(Arc::new(FilterPushDown::new()));
         rules.push(Arc::new(LimitPushDown::new()));
         rules.push(Arc::new(SingleDistinctToGroupBy::new()));
